pyproject.toml
[project]
name = "sparc-enhanced"
version = "2.0.0"
description = "SPARC Enhanced Autonomous Development Framework with Memory Intelligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI dependencies
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pydantic[email]==2.5.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    
    # Testing dependencies
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "bandit[toml]==1.7.5",
    
    # SPARC Memory Intelligence System Dependencies
    "supabase>=2.0.0",
    "qdrant-client>=1.7.0",
    "rich>=13.0.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.3.0",
    "numpy>=1.24.0",
    
    # Additional AI/ML dependencies for enhanced features
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    
    # Data processing and utilities
    "pandas>=2.0.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    "requests>=2.28.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []